Be/feat/aiinterview

See merge request s15-webmobile1-sub1/S15P11D202!40

+++ be/src/main/java/com/aitserver/aiInterview/entity/AiComprehensiveReport.java
@@ -31,27 +31,27 @@ public class AiComprehensiveReport {
     private String content;
 
     @Column(name = "eye_contact_score", nullable = false)
-    private Integer eyeContactScore;
+    private Double eyeContactScore;
 
     @Column(name = "face_score", nullable = false)
-    private Integer faceScore;
+    private Double faceScore;
 
     @Column(name = "voice_score", nullable = false)
-    private Integer voiceScore;
+    private Double voiceScore;
 
     @Column(name = "qna_score", nullable = false)
-    private Integer qnaScore;
+    private Double qnaScore;
 
     @Column(name = "sentence_score", nullable = false)
-    private Integer sentenceScore;
+    private Double sentenceScore;
 
     @CreationTimestamp
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
 
     @Builder
-    public AiComprehensiveReport(Long aiInterviewId, String content, Integer eyeContactScore,
-                                 Integer faceScore, Integer voiceScore, Integer qnaScore, Integer sentenceScore) {
+    public AiComprehensiveReport(Long aiInterviewId, String content, Double eyeContactScore,
+                                 Double faceScore, Double voiceScore, Double qnaScore, Double sentenceScore) {
         this.aiInterviewId = aiInterviewId;
         this.content = content;
         this.eyeContactScore = eyeContactScore;
